+++ laravel/database/migrations/2024_08_13_104550_create_prices_table.php
@@ -21,9 +21,9 @@ return new class extends Migration
             $table->string('technical_drawing');
             $table->string('product_image');
             $table->text('description')->nullable();
-            $table->int('order');
+            $table->integer('order');
             $table->boolean('is_deleted')->default(false);
-            $table->int('company_id');
+            $table->unsignedBigInteger('company_id');
             $table->timestamps();
 
             $table->foreign('company_id')->references('id')->on('companies')->onDelete('restrict');
